Move ProgramCache to device/mod.rs
